seperate content layer export and import
