--- fallctf.com/src/components/Footer.astro
@@ -1,3 +1,0 @@
-<div>
-  footer
-</div>
