+++ fallctf.com/src/content/config.ts
@@ -1,3 +1,3 @@
-import { events } from '~/content/config';
+import { events } from '$/utils/content';
 
 export const collections = { events }
