+++ fallctf.com/tsconfig.json
@@ -8,7 +8,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"],
-      "~/*": ["../_global/*"]
+      "$/*": ["../_global/*"]
     }
   }
 }
