+++ sigpwny.com/src/layouts/Base.astro
@@ -1,10 +1,10 @@
 ---
 import '@/styles/main.css';
 import '@/styles/md.css';
-import '~/styles/prism-one-dark.css';
+import '$/styles/prism-one-dark.css';
 
-import type { HeadProps } from '~/components/BaseHead.astro';
-import BaseHead from '~/components/BaseHead.astro';
+import type { HeadProps } from '$/components/BaseHead.astro';
+import BaseHead from '$/components/BaseHead.astro';
 import Header from '@/components/Header.astro';
 import Footer from '@/components/Footer.astro';
